requirements.txt
fastapi[standard]
uvicorn[standard]
automata-lib[visual]